feat: add split amount

+++ src/endpoints/quest_boost/get_claim_params.rs
@@ -34,7 +34,7 @@ pub async fn handler(
 
     let boost: Document = res.unwrap();
     let num_of_winners = boost.get("num_of_winners").unwrap().as_i32().unwrap();
-    let amount = boost.get("amount").unwrap().as_i32().unwrap() / num_of_winners;
+    let amount = boost.get("amount").unwrap().as_i32().unwrap() as u32 / num_of_winners as u32;
     let token = boost.get("token").unwrap().as_str().unwrap();
     let address = boost.get("winner").unwrap().as_str().unwrap();
